+++ docs/CHANGELOG.md
@@ -3,6 +3,11 @@
 All notable changes to Data Machine will be documented in this file. Also viewable at: 
 
 
+## Unreleased
+
+- CLI: Added --set-prompt option to flows update command for updating handler step prompts via WP-CLI
+- CLI: flows get now shows prompt preview column (truncated to 50 chars)
+
 ## [0.21.1] - 2026-02-05
 
 ### Added